slowly but surely i'm gonna make it

+++ include_libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: fcatinau <fcatinau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/09 15:35:06 by fcatinau          #+#    #+#             */
-/*   Updated: 2021/01/30 09:50:56 by fcatinau         ###   ########.fr       */
+/*   Updated: 2021/01/31 16:05:12 by fcatinau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int		ft_strlen(char *c)
+int			ft_strlen(char *c)
 {
-	int i;
+	int	i;
 
 	i = 0;
 	while (c[i])
@@ -22,9 +22,9 @@ int		ft_strlen(char *c)
 	return (i);
 }
 
-int		ft_strlennb(char *s, size_t *pos)
+int			ft_strlennb(char *s, size_t *pos)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (ft_verif_isdigit(s[*pos]))
